+++ src/algorithms/properties/mod.rs
@@ -29,7 +29,7 @@ pub use girth::girth;
 pub use harmonic::harmonic_centrality;
 pub use is_simple::is_simple;
 pub use knn::avg_nearest_neighbor_degree;
-pub use multiplicity::{has_loop, has_multiple};
+pub use multiplicity::{has_loop, has_multiple, is_loop, is_multiple};
 pub use pagerank::pagerank;
 pub use reciprocity::reciprocity;
 pub use triangles::{count_triangles, transitivity_local_undirected, transitivity_undirected};